Make form control elements reusable
We started with a single `Input` component. Now we
want some of its functionality to be shared.
